Finalized notebook to modularize functions/ ran pytests
